Fixed the Image caption selection
There was a bug in the Image Caption Field Selector. One of the class name in the class list of the element was randomly generated. Hence, when the field is looked for with absolute value, it is not found. That field is now replaced by looking for only the known class names